Move inner Get[] evaluation to mathics/eval/files_io/files.py
